requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
numpy==2.0.2
python-dotenv==1.0.1
openai==1.58.1
pymongo[srv]==4.10.1
